+++ tests/test_providers.py
@@ -106,7 +106,7 @@ class ProviderTests(type):
 
         def search_strings(self, mode):
             _search_strings = {
-                'RSS': self.provider.cache.search_params['RSS'],
+                'RSS': self.provider.cache.search_strings['RSS'],
                 'Episode': self.provider._get_episode_search_strings(self.ep)[0]['Episode'],
                 'Season': self.provider._get_season_search_strings(self.ep)[0]['Season']
             }